Changes to fix duplication of text
I made changes to these pages as text was duplicated. You can see the problem on these pages on the Wilde site. But I am not sure I have kept all the right tags. Does this look okay to you? If so, I will change the rest of the report pages in the same way.